requirements.txt
APScheduler==3.6.3
astroid==2.12.12
async-generator==1.10
attrs==22.1.0
black==22.10.0
cachetools==4.2.2
certifi==2022.6.15
click==8.1.3
dill==0.3.5.1
h11==0.13.0
humanize==4.4.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
outcome==1.2.0
pathspec==0.10.1
platformdirs==2.5.2
pylint==2.15.5
PySocks==1.7.1
python-telegram-bot==13.14
pytz==2022.5
pytz-deprecation-shim==0.1.0.post0
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.11.5
tornado==6.1
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
wrapt==1.14.1
wsproto==1.2.0
